requirements.txt
streamlit==1.31.0
pytesseract==0.3.10
Pillow==10.2.0
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
